@@ -10,23 +10,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "^12.2.16",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
-    "@po-ui/ng-components": "5.22.7",
-    "@po-ui/ng-templates": "^6.13.1",
-    "@totvs/po-totvs-corp-theme": "^2.0.0-beta.1",
-    "protheus-lib-core": "^8.2.0",
-    "rxjs": "~6.6.0",
-    "subsink": "^1.0.2",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "14.1.2",
+    "@angular/cdk": "14.1.2",
+    "@angular/common": "14.1.2",
+    "@angular/compiler": "14.1.2",
+    "@angular/core": "14.1.2",
+    "@angular/flex-layout": "14.0.0-beta.40",
+    "@angular/forms": "14.1.2",
+    "@angular/platform-browser": "14.1.2",
+    "@angular/platform-browser-dynamic": "14.1.2",
+    "@angular/router": "14.1.2",
+    "@po-ui/ng-components": "14.2.1",
+    "@po-ui/ng-templates": "14.2.1",
+    "@po-ui/style": "14.2.1",
+    "@totvs/po-totvs-corp-theme": "^1.11.0",
+    "null": "2.0.0",
+    "rxjs": "7.5.6",
+    "tslib": "2.4.0",
+    "zone.js": "0.11.8"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~12.2.17",
